\name{ITRSelect-package}
\alias{ITRSelect-package}
\alias{ITRSelect}
\docType{package}
\title{
	Variable Selection for Optimal Individualized Dynamic Treatment Regime
}
\description{
	Sequential advantage selection (SAS) and penalized A-learning (PAL) methods
	are implement for selecting important variables involved in optimal individualized 
	(dynamic) treatment regime in both single-stage or multi-stage studies.
}
\details{
\tabular{ll}{
Package: \tab ITRSelect\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2018-04-19\cr
License: \tab GPL-2\cr
}

}
\author{
Chengchun Shi, Ailin Fan, Rui Song, Wenbin Lu

Maintainer: Chengchun Shi <cshi4@ncsu.edu>
}
\references{
Shi, C. and Fan, A. and Song, R. and Lu, W. (2018) High-Dimensional A-Learing for Optimal
Dynamic Treatment Regimes. \emph{Annals of Statistics,} \bold{ 46:} 925-957. 

Fan, A. and Lu, W. and Song, R. (2016) Sequential Advantage Selection for Optimal Treatment 
Regime. \emph{Annals of Applied Statistics,} \bold{ 10:} 32-53. 

Shi, C. and Song, R. and Lu, W. (2018) Concordance and Value Information Criteria for
Optimal Treatment Decision. \emph{Under review}. 	
}